chore(common): CATALYST-1444 move hooks over into apps/storefront
