+++ apps/storefront/package.json
@@ -22,7 +22,6 @@
   },
   "dependencies": {
     "@b3/global-b3": "*",
-    "@b3/hooks": "*",
     "@emotion/cache": "^11.9.3",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.9.3",
