+++ apps/storefront/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useB2BCallback';
 export { default as useB3AppOpen } from './useB3AppOpen';
 export { default as useHideGoogleCustomerReviews } from './dom/useHideGoogleCustomerReviews';
 export { default as useRegisteredbctob2b } from './dom/useRegisteredbctob2b';
@@ -13,3 +14,4 @@ export { default as useSetOpen } from './useSetOpen';
 export { default as useSort } from './useSort';
 export { default as useTableRef } from './useTableRef';
 export * from './useVerifyPermission';
+export { default as useWindowSize } from './useWindowSize';
