+++ apps/storefront/src/pages/Login/useLogout.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
-import { dispatchEvent } from '@b3/hooks';
 
+import { dispatchEvent } from '@/hooks';
 import { endUserMasqueradingCompany, superAdminEndMasquerade } from '@/shared/service/b2b';
 import { bcLogoutLogin } from '@/shared/service/bc';
 import { clearMasqueradeCompany, useAppDispatch, useAppSelector } from '@/store';
