--- packages/hooks/.lintstagedrc.json
@@ -1,4 +1,0 @@
-{
-  "*.{ts,tsx,js}": "eslint --fix",
-  "*.json": "prettier --write"
-}
