--- packages/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useB2BCallback';
-export { default as useWindowSize } from './useWindowSize';
