--- packages/hooks/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@b3/hooks",
-  "version": "0.0.0",
-  "main": "index.ts",
-  "private": true,
-  "prettier": "@bigcommerce/eslint-config/prettier",
-  "scripts": {
-    "lint": "eslint . --ext .ts,.tsx,.js --max-warnings 0 .",
-    "format": "yarn lint --fix"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@b3/global-b3": "*"
-  },
-  "devDependencies": {
-    "@b3/tsconfig": "*",
-    "@types/react": "^18.3.3",
-    "eslint": "^8.2.0",
-    "eslint-config-bc": "*",
-    "typescript": "^5.9.2"
-  }
-}
